requirements.txt
python-telegram-bot==20.5
cachetools==5.3.3
requests==2.31.0